+++ components/src/lib/Button/Button.type.ts
@@ -1,0 +1,1 @@
+export type ButtonColorType = 'dark' | 'light';
