+++ components/src/lib/index.ts
@@ -1,2 +1,2 @@
-export { default as Button } from './Button';
+export { default as Button } from './Button/Button';
 export { default as Modal } from './Modal/Modal';
