+++ components/src/types/common.type.ts
@@ -1,0 +1,1 @@
+export type Size = 'small' | 'medium' | 'large';
